Overhaul to design and behaviours, moving to more MVVM like design
Lots of changes here, perhaps too many to recall. I'm still learning the best way to put the app together and have come to find MVVM to be a pretty cool design pattern, will be trying to adopt this going forward. This is the most functional the app has been, not mention the fastest too.